Added fixes for issues mentioned in code review. Separated stage_1,2 and stage 3 logic.
