+++ stage_1,2_oma_single_run_parsing.py
@@ -4,7 +4,7 @@ from downloader import Downloader
 from db_configurator import get_config_string
 
 
-def descrete_parsing():
+def main():
     logging.basicConfig(filename='oma_parsing.log', level=logging.INFO,
                         format='%(asctime)s %(message)s',
                         datefmt='%m/%d/%Y %I:%M:%S %p')
@@ -19,11 +19,11 @@ def descrete_parsing():
         download.parse_main_catalog_page_single_run()
         logging.info('Stage_1 parsing is finished')
     logging.info('Stage_2 parsing started')
-    if not download.check_if_stage2_parsing_is_complete():
+    while not download.check_if_stage2_parsing_is_complete():
         logging.info('Successful check of incompletion of stage_2 parsing')
         download.parse_product_url_from_subcats_lvl2(number_of_subcats=1)
     logging.info('Stage_2 parsing finished')
-    if not download.check_if_stage3_parsing_is_complete():
-        entry_id = download.parse_products_parameters(number_of_products_to_parse=1)
-    logging.info('Parsing is complete.')
-    return True
+    quit()
+
+if __name__ == '__main__':
+    main()
